the entire backend has been built and tested.
cleaned code, added comments, better naming, and reasonable structure of the codebase.

+++ lib/conversions.py
@@ -1,3 +1,5 @@
+"""methods to convert images into pdf or cbz format"""
+
 import zipfile
 from pathlib import Path
 
@@ -5,12 +7,18 @@ import img2pdf
 
 
 def convert_images_to_pdf(image_dir: Path, pdf_path: Path):
+    if pdf_path.exists():
+        return
+    # sorts the image just in case.
     images = sorted(image_dir.glob("*.png"))
     with open(pdf_path, "wb") as f:
         f.write(img2pdf.convert([str(image) for image in images]))
 
 
 def convert_images_to_cbz(image_dir: Path, cbz_path: Path):
+    if cbz_path.exists():
+        return
+    # sorts the image just in case.
     images = sorted(image_dir.glob("*.png"))
     with zipfile.ZipFile(cbz_path, "w") as cbz:
         for image in images:
